330: rename data.js -> backwardsFunnelData.js
